requirements.txt
streamlit==1.37.1
pymupdf==1.24.9
reportlab==4.2.2
pillow==10.4.0
pyzipper==0.3.6